LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-getElementsByTagName-expected.txt


PASS Interfaces 
PASS Caching is allowed 
FAIL Shouldn't be able to set unsigned properties on a HTMLCollection (non-strict mode) assert_equals: expected (undefined) undefined but got (string) "foopy"
FAIL Shouldn't be able to set unsigned properties on a HTMLCollection (strict mode) assert_throws: function "function () {
      "use strict";
      l[5] = "foopy"
    }" did not throw
PASS Should be able to set expando shadowing a proto prop (item) 
PASS Should be able to set expando shadowing a proto prop (namedItem) 
FAIL hasOwnProperty, getOwnPropertyDescriptor, getOwnPropertyNames assert_true: desc.configurable expected true got false
PASS HTML element with uppercase tagName never matches in HTML Documents 
PASS Element in non-HTML namespace, no prefix, lowercase name 
PASS Element in non-HTML namespace, no prefix, uppercase name 
FAIL Element in non-HTML namespace, prefix, lowercase name assert_array_equals: lengths differ, expected 0 got 1
FAIL Element in non-HTML namespace, prefix, uppercase name assert_array_equals: lengths differ, expected 0 got 1
PASS Element in HTML namespace, no prefix, non-ascii characters in name 
PASS Element in non-HTML namespace, non-ascii characters in name 
PASS Element in HTML namespace, prefix, non-ascii characters in name 
PASS Element in non-HTML namespace, prefix, non-ascii characters in name 
PASS getElementsByTagName('*') 
